+++ packages/scheduler/src/utils/execution.ts
@@ -27,16 +27,19 @@ export function computePotentialExecutionTime<ExecutionDataType>(
 		case 'fixed-round':
 			return timing.expectedTime;
 		case 'delta-time':
-			return (
-				timing.delta +
-				(execution.priorTransactionConfirmation
+			const startTime = execution.priorTransactionConfirmation
+				? execution.priorTransactionConfirmation?.startTime
 					? execution.priorTransactionConfirmation?.startTime
-						? execution.priorTransactionConfirmation?.startTime
-						: execution.priorTransactionConfirmation?.blockTime
-					: state?.startTimeToCountFrom
-						? state.startTimeToCountFrom
-						: timing.startTransaction.broadcastTime)
-			);
+					: execution.priorTransactionConfirmation?.blockTime
+				: state?.startTimeToCountFrom
+					? state.startTimeToCountFrom
+					: timing.startTransaction.broadcastTime;
+
+			const startTimePlusDelta = startTime + timing.delta;
+			if (timing.targetTimeUnlessHigherDelta && timing.targetTimeUnlessHigherDelta > startTimePlusDelta) {
+				return timing.targetTimeUnlessHigherDelta;
+			}
+			return startTimePlusDelta;
 	}
 
 	throw new Error(`execution timing type must be "fixed-time" | "fixed-round" | "delta-time"`);