+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # fuzd-server
 
+## 0.9.6
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-scheduler@0.9.6
+  - fuzd-tlock-decrypter@0.9.6
+
 ## 0.9.5
 
 ### Patch Changes